Improved Optimization Process (Still Broken)
The dead code optimization will still break under certain circumstances.

It will produce a warning about your program going out of bounds even if it doesn't.

HOWEVER, this branch can be used to compile programs, but if the dead code optimization fails it will move on and only perform the consecutive statement optimization.